Move test_not_rematerialize_matmul to test_nvfuser_remat.py
